% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cuba.R
\name{read_cuba}
\alias{read_cuba}
\title{Read in *.grib2 data}
\usage{
read_cuba(grib2)
}
\arguments{
\item{grib2}{Chr Canonical Filename}
}
\value{
Tibble
}
\description{
Read in *.grib2 data
}
\examples{
path <- file.path("inst", "extdata")
files <- list.files(path, full.names = TRUE, pattern = "\\\\.grb2")
grib2 <- files[1]
}
